option display renaming of interfaces
