+++ DEHPCommon/UserInterfaces/ViewModels/Interfaces/IObjectBrowserViewModel.cs
@@ -1,6 +1,6 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
 // <copyright file="IObjectBrowserViewModel.cs" company="RHEA System S.A.">
-//    Copyright (c) 2020-2020 RHEA System S.A.
+//    Copyright (c) 2020-2021 RHEA System S.A.
 // 
 //    Author: Sam Gerené, Alex Vorobiev, Alexander van Delft, Nathanael Smiechowski.
 // 
@@ -26,16 +26,85 @@ namespace DEHPCommon.UserInterfaces.ViewModels.Interfaces
 {
     using System;
 
-    using CDP4Common.CommonData;
-
     using DEHPCommon.Services.ObjectBrowserTreeSelectorService;
 
     using ReactiveUI;
 
     /// <summary>
-    /// Interface definition for the <see cref="ObjectBrowserViewModel"/>
+    /// Interface definition for <see cref="ObjectBrowserBaseViewModel"/>
     /// </summary>
-    public interface IObjectBrowserViewModel : IObjectBrowserBaseViewModel
+    public interface IObjectBrowserViewModel
     {
+        /// <summary>
+        /// Gets or sets a value indicating whether the browser is busy
+        /// </summary>
+        bool? IsBusy { get; set; }
+
+        /// <summary>
+        /// Gets or sets the selected thing
+        /// </summary>
+        object SelectedThing { get; set; }
+
+        /// <summary>
+        /// Gets or sets the selected things collection
+        /// </summary>
+        ReactiveList<object> SelectedThings { get; set; }
+
+        /// <summary>
+        /// Gets the collection of <see cref="BrowserViewModelBase"/> to be displayed in the tree
+        /// </summary>
+        ReactiveList<BrowserViewModelBase> Things { get; }
+
+        /// <summary>
+        /// Gets the Context Menu for the implementing view model
+        /// </summary>
+        ReactiveList<ContextMenuItemViewModel> ContextMenu { get; }
+
+        /// <summary>
+        /// Gets the command that allows to map the selected things
+        /// </summary>
+        ReactiveCommand<object> MapCommand { get; set; }
+
+        /// <summary>
+        /// Gets or sets the <see cref="IObservable{T}"/> of <see cref="bool"/> that is bound to the <see cref="MapCommand"/> <see cref="ReactiveCommand{T}.CanExecute"/> property
+        /// </summary>
+        /// <remarks>This observable is intended to be Merged with another observable</remarks>
+        IObservable<bool> CanMap { get; set; }
+
+        /// <summary>
+        /// Gets the Caption of the control
+        /// </summary>
+        string Caption { get; }
+
+        /// <summary>
+        /// Gets the tooltip of the control
+        /// </summary>
+        string ToolTip { get; }
+
+        /// <summary>
+        /// Updates the tree
+        /// </summary>
+        /// <param name="shouldReset">A value indicating whether the tree should remove the element in preview</param>
+        void UpdateTree(bool shouldReset);
+
+        /// <summary>
+        /// Reloads the the trees elements
+        /// </summary>
+        void Reload();
+
+        /// <summary>
+        /// Adds to the <see cref="ObjectBrowserBaseViewModel.Things"/> collection the specified by <see cref="IObjectBrowserTreeSelectorService"/> trees
+        /// </summary>
+        void BuildTrees();
+
+        /// <summary>
+        /// Populate the context menu for the implementing view model
+        /// </summary>
+        void PopulateContextMenu();
+
+        /// <summary>
+        /// Dispose of this <see cref="ObjectBrowserBaseViewModel"/>
+        /// </summary>
+        void Dispose();
     }
 }
